get next link working and write comments for monday
