+++ src/compounds/Item/item.scss
@@ -10,42 +10,15 @@
   width: 250px;
 }
 
-.item-details {
-  font-family: 'Inter', 'Helvetica Neue', Helvetica, Arial, sans-serif;
-  padding-right: 15px;
-}
-
-.item-description {
-  font-size: 16px;
-  font-weight: 300;
-}
-
-.item-name {
-  font-size: 20px;
-  font-weight: 500;
-}
-
-.item-image {
-  width: 100%;
-}
-
 .item-link:hover {
   color: #A9A9A9;
 }
 
-.item-name-button {
-  cursor: pointer;
+.item-vertical + .item-vertical {
+  margin-left: 5px;
 }
 
-.item-button-horizontal {
-  width: 165px;
-}
 
-.item-button-vertical {
-  margin: 15px;
-  display: flex;
-  flex-direction: row-reverse;
-}
 
 /************ vertical items ************/
 
@@ -57,24 +30,6 @@
   height: 350px;
 }
 
-.item-vertical > .item-image {
-  height: 150px;
-  object-fit: cover;
-}
-
-.item-vertical + .item-vertical {
-  margin-left: 5px;
-}
-
-.item-button-vertical {
-  margin: 15px;
-  display: flex;
-  flex-direction: row-reverse;
-}
-
-.item-options-vertical > .item-details {
-  padding-left: 15px;
-}
 
 /************ horizontal items ************/
 
@@ -83,19 +38,6 @@
   width: auto;
   height: 175px;
   margin-bottom: 30px;
-}
-
-.item-horizontal.item-details {
-  width: 81.5%;
-}
-
-.item-image-horizontal {
-  width: 240px;
-  border-right: 1px solid;
-}
-
-.item-button-horizontal {
-  width: 165px;
 }
 
 .item-options-horizontal {
