+++ src/compounds/ItemGroup/ItemGroup.stories.jsx
@@ -9,9 +9,26 @@ export default {
 
 const Template = (args) => <ItemGroup {...args} />
 
-export const Default = Template.bind({})
-Default.args = {
+export const Vertical = Template.bind({})
+Vertical.args = {
   items: items.slice(0, 3),
+  buttonLink: '',
+  buttonProps: {
+    backgroundColor: '#A9A9A9',
+    label: 'Request this item',
+  },
+  orientation: 'vertical'
+}
+
+export const Horizontal = Template.bind({})
+Horizontal.args = {
+  items: items.slice(0, 3),
+  buttonLink: '',
+  buttonProps: {
+    backgroundColor: '#A9A9A9',
+    label: 'Request this item',
+  },
+  orientation: 'horizontal'
 }
 
 export const isLoading = Template.bind({})
